backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.31.0
sqlalchemy==2.0.35
beautifulsoup4==4.12.3
requests==2.32.3
pydantic==2.9.2
numpy>=1.24.0
langchain-core>=0.3.15
langchain-community>=0.3.5
langchain-google-genai>=2.0.4
python-dotenv==1.0.1